Document what we did so far
So far, I've created the public repo with truncated history, and done
one new sync of master to the public side.

I note that each time this happens, all the actual branching structure
that is present in the real master is hidden in the public one.  This
might actually be desirable, or it could be a huge problem.  I have no
idea.