Add info to README
This adds a note to the README pointing out the limited support of the
GitHub side of the repo, and discouraging pull requests.  I have also
disabled the project's issue tracker lest people start using it
instead of the mailing list.

(Of course, there IS no mailing list, but this is what I intend to do
for Xyce when we have a Xyce repo on GitHub).