Fix early discussion after-the-fact
Early on, I made a mistake on where to start the public history, and
the manifesto started describing what I should have done, then showed
what I did do --- I eventually realized the error and commented on it,
but forgot to go back and make the earlier text correct.  As it stood,
I said "we're going to do X" and then I actually showed doing X+deltaX
instead.

Fix it.